Platform-gate Unix shell-outs and wire Windows CI/build (🎯T22)
Split the Unix-only lsof/ps/vm_stat shell-outs out of internal/store/store.go
into store_unix.go (//go:build !windows) and add store_windows.go stubs
that degrade gracefully. Live-session discovery and per-PID process
metrics become no-ops on Windows; transcript indexing and all query
tools work identically. mnemo_whatsup returns an empty session list
on Windows rather than crashing.

Refactor the inline ps -o call inside Whatsup into a runPsMetrics seam
mirroring the existing runPsEnv seam, and extract parsePsMetricsOutput
as a pure parser alongside parsePsEnvOutput. vm_stat parsing likewise
moves to parseVMStatOutput. All three parsers stay in store.go so
they're exercised on every platform.

Add .github/workflows/ci.yml running go test on ubuntu-latest and
windows-latest with CGO_ENABLED=1 (both runners ship with MinGW-w64 /
gcc). Extend release.yml with a windows-latest amd64 entry that
static-links the MinGW runtime via -extldflags "-static" and packages
the binary as a .zip via 7z. Document the MinGW-w64 prereq for
Windows source builds in README.

Depends on a claudia release containing the Flock → LockFileEx split
(flock_unix.go / flock_windows.go on the claudia side). CI on Windows
will stay red until mnemo's go.mod pins that release.
