+++ .github/workflows/release.yml
@@ -15,13 +15,32 @@ jobs:
           - os: macos-latest
             goos: darwin
             goarch: arm64
+            ext: ""
+            archive: tar.gz
           - os: ubuntu-latest
             goos: linux
             goarch: amd64
+            ext: ""
+            archive: tar.gz
           - os: ubuntu-24.04-arm
             goos: linux
             goarch: arm64
+            ext: ""
+            archive: tar.gz
+          - os: windows-latest
+            goos: windows
+            goarch: amd64
+            ext: ".exe"
+            archive: zip
+            # Static-link the MinGW runtime so end users do not need
+            # MinGW-w64 DLLs on their PATH. 🎯T22 AC: "Release binary
+            # on Windows is statically linked (or bundles required
+            # DLLs)".
+            ldflags_extra: '-extldflags "-static"'
     runs-on: ${{ matrix.os }}
+    defaults:
+      run:
+        shell: bash
     steps:
       - uses: actions/checkout@v4
         with:
@@ -33,8 +52,9 @@ jobs:
           path: sqldeep
 
       - name: Build sqldeep
+        working-directory: sqldeep
         run: |
-          cd sqldeep && mkdir -p build
+          mkdir -p build
           c++ -std=c++20 -Wall -Wextra -Wpedantic -Idist -Ivendor/include -c dist/sqldeep.cpp -o build/sqldeep.o
           cc -w -Idist -Ivendor/include -c dist/sqldeep_xml.c -o build/sqldeep_xml.o
           ar rcs build/libsqldeep.a build/sqldeep.o
@@ -46,13 +66,23 @@ jobs:
       - name: Build
         env:
           CGO_ENABLED: "1"
+          GOOS: ${{ matrix.goos }}
+          GOARCH: ${{ matrix.goarch }}
+          LDFLAGS_EXTRA: ${{ matrix.ldflags_extra }}
         working-directory: mnemo
         run: |
           VERSION="${GITHUB_REF_NAME#v}"
           go build -tags "sqlite_fts5" \
-            -ldflags "-s -w" \
-            -o mnemo .
-          tar czf "mnemo-${VERSION}-${{ matrix.goos }}-${{ matrix.goarch }}.tar.gz" mnemo
+            -ldflags "-s -w ${LDFLAGS_EXTRA}" \
+            -o "mnemo${{ matrix.ext }}" .
+          case "${{ matrix.archive }}" in
+            tar.gz)
+              tar czf "mnemo-${VERSION}-${{ matrix.goos }}-${{ matrix.goarch }}.tar.gz" "mnemo${{ matrix.ext }}"
+              ;;
+            zip)
+              7z a "mnemo-${VERSION}-${{ matrix.goos }}-${{ matrix.goarch }}.zip" "mnemo${{ matrix.ext }}"
+              ;;
+          esac
 
       - name: Upload release artifact
         env:
@@ -61,7 +91,7 @@ jobs:
         run: |
           VERSION="${GITHUB_REF_NAME#v}"
           gh release upload "$GITHUB_REF_NAME" \
-            "mnemo-${VERSION}-${{ matrix.goos }}-${{ matrix.goarch }}.tar.gz"
+            "mnemo-${VERSION}-${{ matrix.goos }}-${{ matrix.goarch }}.${{ matrix.archive }}"
 
   test:
     runs-on: ubuntu-latest
